refactor(plan-lifecycle): migrate runner.ts to plan-lifecycle facade
Replace all direct imports from plan-detection, frontmatter, receipt,
global-state, and pipeline-state in runner.ts and its 7 test files with
imports from the unified plan-lifecycle facade.
